@@ -1,15 +1,14 @@
 package com.tiviacz.travelersbackpack.capability;
 
 import com.tiviacz.travelersbackpack.TravelersBackpack;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.nbt.INBT;
-import net.minecraft.util.Direction;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.CapabilityInject;
-import net.minecraftforge.common.capabilities.CapabilityManager;
-import net.minecraftforge.common.capabilities.ICapabilityProvider;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.common.capabilities.*;
+import net.minecraftforge.common.util.LazyOptional;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 public class TravelersBackpackCapability
 {
@@ -22,12 +21,15 @@ public class TravelersBackpackCapability
 
     public static void register()
     {
-        CapabilityManager.INSTANCE.register(ITravelersBackpack.class, new Capability.IStorage<ITravelersBackpack>()
+        CapabilityManager.INSTANCE.register(ITravelersBackpack.class);
+    }
+
+    /*, new Capability.IStorage<ITravelersBackpack>()
         {
             @Override
-            public INBT writeNBT(final Capability<ITravelersBackpack> capability, final ITravelersBackpack instance, final Direction side)
+            public Tag writeNBT(final Capability<ITravelersBackpack> capability, final ITravelersBackpack instance, final Direction side)
             {
-                CompoundNBT compound = new CompoundNBT();
+                CompoundTag compound = new CompoundTag();
 
                 if(instance.hasWearable())
                 {
@@ -43,17 +45,48 @@ public class TravelersBackpackCapability
             }
 
             @Override
-            public void readNBT(final Capability<ITravelersBackpack> capability, final ITravelersBackpack instance, final Direction side, final INBT nbt)
+            public void readNBT(final Capability<ITravelersBackpack> capability, final ITravelersBackpack instance, final Direction side, final Tag nbt)
             {
-                CompoundNBT stackCompound = (CompoundNBT)nbt;
+                CompoundTag stackCompound = (CompoundTag)nbt;
                 ItemStack wearable = ItemStack.of(stackCompound);
                 instance.setWearable(wearable);
             }
         }, () -> new TravelersBackpackWearable(null));
-    }
+    } */
 
     public static ICapabilityProvider createProvider(final ITravelersBackpack backpack)
     {
-        return new SerializableCapabilityProvider<>(TRAVELERS_BACKPACK_CAPABILITY, DEFAULT_FACING, backpack);
+        return new Provider(backpack);
+    }
+
+    public static class Provider implements ICapabilitySerializable<CompoundTag>
+    {
+        final ITravelersBackpack backpack;
+        final LazyOptional<ITravelersBackpack> optional;
+
+        public Provider(final ITravelersBackpack backpack)
+        {
+            this.backpack = backpack;
+            this.optional = LazyOptional.of(() -> backpack);
+        }
+
+        @Nonnull
+        @Override
+        public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side)
+        {
+            return TravelersBackpackCapability.TRAVELERS_BACKPACK_CAPABILITY.orEmpty(cap, this.optional);
+        }
+
+        @Override
+        public CompoundTag serializeNBT()
+        {
+            return backpack.saveTag();
+        }
+
+        @Override
+        public void deserializeNBT(CompoundTag nbt)
+        {
+            backpack.loadTag(nbt);
+        }
     }
 }
