@@ -4,20 +4,20 @@ import com.tiviacz.travelersbackpack.capability.CapabilityUtils;
 import com.tiviacz.travelersbackpack.capability.ITravelersBackpack;
 import com.tiviacz.travelersbackpack.config.TravelersBackpackConfig;
 import com.tiviacz.travelersbackpack.fluids.EffectFluidRegistry;
-import com.tiviacz.travelersbackpack.inventory.TravelersBackpackInventory;
-import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackItemContainer;
+import com.tiviacz.travelersbackpack.inventory.TravelersBackpackContainer;
+import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackItemMenu;
 import com.tiviacz.travelersbackpack.items.HoseItem;
-import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackTileEntity;
+import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackBlockEntity;
 import com.tiviacz.travelersbackpack.util.FluidUtils;
 import com.tiviacz.travelersbackpack.util.Reference;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.Hand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
+import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fluids.capability.IFluidHandler;
@@ -26,18 +26,18 @@ import net.minecraftforge.items.ItemStackHandler;
 
 public class ServerActions
 {
-    public static void cycleTool(PlayerEntity player, double scrollDelta)
+    public static void cycleTool(Player player, double scrollDelta)
     {
         if(CapabilityUtils.isWearingBackpack(player))
         {
-            TravelersBackpackInventory inventory = CapabilityUtils.getBackpackInv(player);
-            ItemStackHandler inv = inventory.getInventory();
+            TravelersBackpackContainer inventory = CapabilityUtils.getBackpackInv(player);
+            ItemStackHandler inv = inventory.getHandler();
             ItemStack heldItem = player.getMainHandItem();
 
             if(!inv.getStackInSlot(Reference.TOOL_UPPER).isEmpty() && inv.getStackInSlot(Reference.TOOL_LOWER).isEmpty() || !inv.getStackInSlot(Reference.TOOL_LOWER).isEmpty() && inv.getStackInSlot(Reference.TOOL_UPPER).isEmpty())
             {
                 boolean isUpperEmpty = inv.getStackInSlot(Reference.TOOL_UPPER).isEmpty();
-                player.setItemInHand(Hand.MAIN_HAND, isUpperEmpty ? inv.getStackInSlot(Reference.TOOL_LOWER) : inv.getStackInSlot(Reference.TOOL_UPPER));
+                player.setItemInHand(InteractionHand.MAIN_HAND, isUpperEmpty ? inv.getStackInSlot(Reference.TOOL_LOWER) : inv.getStackInSlot(Reference.TOOL_UPPER));
                 inv.setStackInSlot(isUpperEmpty ? Reference.TOOL_LOWER : Reference.TOOL_UPPER, heldItem);
             }
 
@@ -45,14 +45,14 @@ public class ServerActions
             {
                 if(scrollDelta < 0)
                 {
-                    player.setItemInHand(Hand.MAIN_HAND, inv.getStackInSlot(Reference.TOOL_UPPER));
+                    player.setItemInHand(InteractionHand.MAIN_HAND, inv.getStackInSlot(Reference.TOOL_UPPER));
                     inv.setStackInSlot(Reference.TOOL_UPPER, inv.getStackInSlot(Reference.TOOL_LOWER));
                     inv.setStackInSlot(Reference.TOOL_LOWER, heldItem);
                 }
 
                 else if(scrollDelta > 0)
                 {
-                    player.setItemInHand(Hand.MAIN_HAND, inv.getStackInSlot(Reference.TOOL_LOWER));
+                    player.setItemInHand(InteractionHand.MAIN_HAND, inv.getStackInSlot(Reference.TOOL_LOWER));
                     inv.setStackInSlot(Reference.TOOL_LOWER, inv.getStackInSlot(Reference.TOOL_UPPER));
                     inv.setStackInSlot(Reference.TOOL_UPPER, heldItem);
                 }
@@ -61,22 +61,22 @@ public class ServerActions
         }
     }
 
-    public static void equipBackpack(PlayerEntity player)
+    public static void equipBackpack(Player player)
     {
         LazyOptional<ITravelersBackpack> cap = CapabilityUtils.getCapability(player);
-        World world = player.level;
+        Level level = player.level;
 
-        if(!world.isClientSide)
+        if(!level.isClientSide)
         {
             if(!cap.map(ITravelersBackpack::hasWearable).orElse(false))
             {
-                if(player.containerMenu instanceof TravelersBackpackItemContainer) player.containerMenu.removed(player);
+                if(player.containerMenu instanceof TravelersBackpackItemMenu) player.containerMenu.removed(player);
 
                 ItemStack stack = player.getMainHandItem().copy();
 
                 cap.ifPresent(inv -> inv.setWearable(stack));
                 player.getMainHandItem().shrink(1);
-                world.playSound(null, player.blockPosition(), SoundEvents.ARMOR_EQUIP_LEATHER, SoundCategory.PLAYERS, 1.0F, (1.0F + (world.random.nextFloat() - world.random.nextFloat()) * 0.2F) * 0.7F);
+                level.playSound(null, player.blockPosition(), SoundEvents.ARMOR_EQUIP_LEATHER, SoundSource.PLAYERS, 1.0F, (1.0F + (level.getRandom().nextFloat() - level.getRandom().nextFloat()) * 0.2F) * 0.7F);
 
                 //Sync
                 CapabilityUtils.synchronise(player);
@@ -86,22 +86,22 @@ public class ServerActions
         }
     }
 
-    public static void unequipBackpack(PlayerEntity player)
+    public static void unequipBackpack(Player player)
     {
         LazyOptional<ITravelersBackpack> cap = CapabilityUtils.getCapability(player);
-        World world = player.level;
+        Level level = player.level;
 
       //  CapabilityUtils.onUnequipped(world, player, cap.getWearable());
 
-        if(!world.isClientSide)
+        if(!level.isClientSide)
         {
-            if(player.containerMenu instanceof TravelersBackpackItemContainer) player.containerMenu.removed(player);
+            if(player.containerMenu instanceof TravelersBackpackItemMenu) player.containerMenu.removed(player);
 
             ItemStack wearable = cap.map(ITravelersBackpack::getWearable).orElse(ItemStack.EMPTY).copy();
 
-            if(!player.inventory.add(wearable))
+            if(!player.getInventory().add(wearable))
             {
-                player.sendMessage(new TranslationTextComponent(Reference.NO_SPACE), player.getUUID());
+                player.sendMessage(new TranslatableComponent(Reference.NO_SPACE), player.getUUID());
                 player.closeContainer();
 
                 return;
@@ -110,7 +110,7 @@ public class ServerActions
             if(cap.map(ITravelersBackpack::hasWearable).orElse(false))
             {
                 cap.ifPresent(ITravelersBackpack::removeWearable);
-                world.playSound(null, player.blockPosition(), SoundEvents.ARMOR_EQUIP_LEATHER, SoundCategory.PLAYERS, 1.05F, (1.0F + (world.random.nextFloat() - world.random.nextFloat()) * 0.2F) * 0.7F);
+                level.playSound(null, player.blockPosition(), SoundEvents.ARMOR_EQUIP_LEATHER, SoundSource.PLAYERS, 1.05F, (1.0F + (level.getRandom().nextFloat() - level.getRandom().nextFloat()) * 0.2F) * 0.7F);
 
                 //Sync
                 CapabilityUtils.synchronise(player);
@@ -120,60 +120,60 @@ public class ServerActions
         }
     }
 
-    public static void toggleSleepingBag(PlayerEntity player, BlockPos pos)
+    public static void toggleSleepingBag(Player player, BlockPos pos)
     {
-        World world = player.level;
+        Level level = player.level;
 
-        if(world.getBlockEntity(pos) instanceof TravelersBackpackTileEntity)
+        if(level.getBlockEntity(pos) instanceof TravelersBackpackBlockEntity)
         {
-            TravelersBackpackTileEntity te = (TravelersBackpackTileEntity)world.getBlockEntity(pos);
+            TravelersBackpackBlockEntity te = (TravelersBackpackBlockEntity)level.getBlockEntity(pos);
 
             if(!te.isSleepingBagDeployed())
             {
-                if(te.deploySleepingBag(world, pos))
+                if(te.deploySleepingBag(level, pos))
                 {
                     player.closeContainer();
                 }
                 else
                 {
-                    player.sendMessage(new TranslationTextComponent(Reference.DEPLOY), player.getUUID());
+                    player.sendMessage(new TranslatableComponent(Reference.DEPLOY), player.getUUID());
                 }
             }
             else
             {
-                te.removeSleepingBag(world);
+                te.removeSleepingBag(level);
             }
             player.closeContainer();
         }
     }
 
-    public static void emptyTank(double tankType, PlayerEntity player, World world)
+    public static void emptyTank(double tankType, Player player, Level level)
     {
-        TravelersBackpackInventory inv = CapabilityUtils.getBackpackInv(player);
+        TravelersBackpackContainer inv = CapabilityUtils.getBackpackInv(player);
         FluidTank tank = tankType == 1D ? inv.getLeftTank() : inv.getRightTank();
-        world.playSound(null, player.blockPosition(), FluidUtils.getFluidEmptySound(tank.getFluid().getFluid()), SoundCategory.BLOCKS, 1.0F, 1.0F);
+        level.playSound(null, player.blockPosition(), FluidUtils.getFluidEmptySound(tank.getFluid().getFluid()), SoundSource.BLOCKS, 1.0F, 1.0F);
         tank.drain(TravelersBackpackConfig.SERVER.tanksCapacity.get(), IFluidHandler.FluidAction.EXECUTE);
         player.closeContainer();
 
         //Sync
         CapabilityUtils.synchronise(player);
         CapabilityUtils.synchroniseToOthers(player);
-        inv.markTankDirty();
+        inv.setTankChanged();
     }
 
-    public static boolean setFluidEffect(World world, PlayerEntity player, FluidTank tank)
+    public static boolean setFluidEffect(Level level, Player player, FluidTank tank)
     {
         FluidStack fluidStack = tank.getFluid();
         boolean done = false;
 
-        if(EffectFluidRegistry.hasFluidEffectAndCanExecute(fluidStack, world, player))
+        if(EffectFluidRegistry.hasFluidEffectAndCanExecute(fluidStack, level, player))
         {
-            done = EffectFluidRegistry.executeFluidEffectsForFluid(fluidStack, player, world);
+            done = EffectFluidRegistry.executeFluidEffectsForFluid(fluidStack, player, level);
         }
         return done;
     }
 
-    public static void switchHoseMode(PlayerEntity player, double scrollDelta)
+    public static void switchHoseMode(Player player, double scrollDelta)
     {
         ItemStack hose = player.getMainHandItem();
 
@@ -207,7 +207,7 @@ public class ServerActions
         }
     }
 
-    public static void toggleHoseTank(PlayerEntity player)
+    public static void toggleHoseTank(Player player)
     {
         ItemStack hose = player.getMainHandItem();
 
