@@ -1,21 +1,22 @@
 package com.tiviacz.travelersbackpack.compat.jei;
 
 import com.tiviacz.travelersbackpack.TravelersBackpack;
-import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackItemContainer;
-import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackTileContainer;
+import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackBlockEntityMenu;
+import com.tiviacz.travelersbackpack.inventory.container.TravelersBackpackItemMenu;
 import mezz.jei.api.IModPlugin;
+import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.constants.VanillaRecipeCategoryUid;
 import mezz.jei.api.registration.IRecipeTransferRegistration;
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
 
-@mezz.jei.api.JeiPlugin
+@JeiPlugin
 public class TravelersBackpackPlugin implements IModPlugin
 {
     @Override
     public void registerRecipeTransferHandlers(IRecipeTransferRegistration registration)
     {
-        registration.addRecipeTransferHandler(TravelersBackpackItemContainer.class, VanillaRecipeCategoryUid.CRAFTING, 1, 9, 1, 90);
-        registration.addRecipeTransferHandler(TravelersBackpackTileContainer.class, VanillaRecipeCategoryUid.CRAFTING, 1, 9, 1, 90);
+        registration.addRecipeTransferHandler(TravelersBackpackItemMenu.class, VanillaRecipeCategoryUid.CRAFTING, 1, 9, 1, 90);
+        registration.addRecipeTransferHandler(TravelersBackpackBlockEntityMenu.class, VanillaRecipeCategoryUid.CRAFTING, 1, 9, 1, 90);
     }
 
     @Override
