@@ -2,8 +2,6 @@ package com.tiviacz.travelersbackpack.handlers;
 
 import com.tiviacz.travelersbackpack.TravelersBackpack;
 import com.tiviacz.travelersbackpack.capability.CapabilityUtils;
-import com.tiviacz.travelersbackpack.client.gui.OverlayScreen;
-import com.tiviacz.travelersbackpack.client.gui.WarningScreen;
 import com.tiviacz.travelersbackpack.config.TravelersBackpackConfig;
 import com.tiviacz.travelersbackpack.inventory.container.slot.ToolSlotItemHandler;
 import com.tiviacz.travelersbackpack.items.HoseItem;
@@ -11,16 +9,12 @@ import com.tiviacz.travelersbackpack.items.TravelersBackpackItem;
 import com.tiviacz.travelersbackpack.network.CycleToolPacket;
 import com.tiviacz.travelersbackpack.network.ScreenPacket;
 import com.tiviacz.travelersbackpack.util.Reference;
+import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.entity.player.ClientPlayerEntity;
-import net.minecraft.client.gui.screen.MainMenuScreen;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.settings.KeyBinding;
-import net.minecraft.item.ItemStack;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.GuiOpenEvent;
 import net.minecraftforge.client.event.InputEvent;
-import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -28,39 +22,27 @@ import net.minecraftforge.fml.common.Mod;
 @Mod.EventBusSubscriber(modid = TravelersBackpack.MODID, value = Dist.CLIENT)
 public class ClientEventHandler
 {
-    @SubscribeEvent
-    public static void onGuiOpen(GuiOpenEvent event)
-    {
-        Screen gui = event.getGui();
-
-        if(TravelersBackpackConfig.CLIENT.displayWarning.get() && gui instanceof MainMenuScreen)
-        {
-            event.setGui(new WarningScreen((MainMenuScreen)gui));
-            TravelersBackpackConfig.CLIENT.displayWarning.set(false);
-        }
-    }
-
-    @SubscribeEvent
-    public static void renderExperienceBar(RenderGameOverlayEvent.Post event)
+   /* @SubscribeEvent
+    public static void renderExperienceBar(RenderGameOverlayEvent.PostLayer event)
     {
         if(TravelersBackpackConfig.CLIENT.overlay.enableOverlay.get())
         {
-            if(event.getType() != RenderGameOverlayEvent.ElementType.HOTBAR || event.isCanceled()) return;
+            if(event.getType() != RenderGameOverlayEvent.ElementType.LAYER || event.isCanceled()) return;
 
             if(CapabilityUtils.isWearingBackpack(Minecraft.getInstance().player))
             {
-                OverlayScreen gui = new OverlayScreen();
-                gui.renderOverlay(event.getMatrixStack());
+                //OverlayScreen gui = new OverlayScreen();
+                //gui.renderOverlay(event.getMatrixStack());
             }
         }
-    }
+    } */
 
     @SubscribeEvent
     public static void clientTickEvent(final TickEvent.ClientTickEvent event)
     {
         if(event.phase != TickEvent.Phase.END) return;
 
-        ClientPlayerEntity player = Minecraft.getInstance().player;
+        LocalPlayer player = Minecraft.getInstance().player;
 
         if(player != null && CapabilityUtils.isWearingBackpack(player))
         {
@@ -77,7 +59,7 @@ public class ClientEventHandler
                 }
             }
 
-            KeyBinding key = ModClientEventHandler.CYCLE_TOOL;
+            KeyMapping key = ModClientEventHandler.CYCLE_TOOL;
 
             if(TravelersBackpackConfig.disableScrollWheel && key.consumeClick())
             {
@@ -109,12 +91,12 @@ public class ClientEventHandler
     public static void mouseWheelDetect(InputEvent.MouseScrollEvent event)
     {
         Minecraft mc = Minecraft.getInstance();
-        KeyBinding key1 = ModClientEventHandler.CYCLE_TOOL;
+        KeyMapping key1 = ModClientEventHandler.CYCLE_TOOL;
         double scrollDelta = event.getScrollDelta();
 
         if(!TravelersBackpackConfig.disableScrollWheel && scrollDelta != 0.0)
         {
-            ClientPlayerEntity player = mc.player;
+            LocalPlayer player = mc.player;
 
             if(player != null && player.isAlive() && key1.isDown())
             {
