@@ -11,25 +11,25 @@ import com.tiviacz.travelersbackpack.init.ModItems;
 import com.tiviacz.travelersbackpack.items.TravelersBackpackItem;
 import com.tiviacz.travelersbackpack.network.SyncBackpackCapabilityClient;
 import com.tiviacz.travelersbackpack.util.BackpackUtils;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.CauldronBlock;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.item.ItemEntity;
-import net.minecraft.entity.merchant.villager.VillagerProfession;
-import net.minecraft.entity.merchant.villager.VillagerTrades;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.MerchantOffer;
-import net.minecraft.loot.LootPool;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.world.GameRules;
-import net.minecraft.world.World;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.npc.VillagerProfession;
+import net.minecraft.world.entity.npc.VillagerTrades;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.trading.MerchantOffer;
+import net.minecraft.world.level.GameRules;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.LayeredCauldronBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.LootTableLoadEvent;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
@@ -41,7 +41,7 @@ import net.minecraftforge.event.village.VillagerTradesEvent;
 import net.minecraftforge.event.world.ExplosionEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.network.PacketDistributor;
+import net.minecraftforge.fmllegacy.network.PacketDistributor;
 
 import javax.annotation.Nullable;
 import java.util.Random;
@@ -52,13 +52,13 @@ public class ForgeEventHandler
     @SubscribeEvent
     public static void playerSetSpawn(PlayerSetSpawnEvent event)
     {
-        World world = event.getPlayer().level;
+        Level level = event.getPlayer().level;
 
         if(event.getNewSpawn() != null)
         {
-            Block block = world.getBlockState(event.getNewSpawn()).getBlock();
+            Block block = level.getBlockState(event.getNewSpawn()).getBlock();
 
-            if(!world.isClientSide && block instanceof SleepingBagBlock && !TravelersBackpackConfig.COMMON.enableSleepingBagSpawnPoint.get())
+            if(!level.isClientSide && block instanceof SleepingBagBlock && !TravelersBackpackConfig.COMMON.enableSleepingBagSpawnPoint.get())
             {
                 event.setCanceled(true);
             }
@@ -70,19 +70,19 @@ public class ForgeEventHandler
     {
         ItemStack stack = event.getItemStack();
 
-        if(event.getWorld().isClientSide || event.getPlayer().isCrouching()) return;
+        if(event.getWorld().isClientSide || event.getPlayer().isShiftKeyDown()) return;
 
         if(stack.getItem() == ModItems.STANDARD_TRAVELERS_BACKPACK.get())
         {
             BlockState blockState = event.getWorld().getBlockState(event.getPos());
 
-            if(BackpackDyeRecipe.hasColor(stack) && blockState.getBlock() instanceof CauldronBlock)
+            if(BackpackDyeRecipe.hasColor(stack) && blockState.getBlock() instanceof LayeredCauldronBlock)
             {
-                if(blockState.getValue(CauldronBlock.LEVEL) > 0)
+                if(blockState.getValue(LayeredCauldronBlock.LEVEL) > 0)
                 {
                     stack.getTag().remove("Color");
-                    ((CauldronBlock)blockState.getBlock()).setWaterLevel(event.getWorld(), event.getPos(), blockState, blockState.getValue(CauldronBlock.LEVEL) - 1);
-                    event.getWorld().playSound(null, event.getPos().getX(), event.getPos().getY(), event.getPos().getY(), SoundEvents.BUCKET_FILL, SoundCategory.BLOCKS, 1.0F, 1.0F);
+                    LayeredCauldronBlock.lowerFillLevel(blockState, event.getWorld(), event.getPos());
+                    event.getWorld().playSound(null, event.getPos().getX(), event.getPos().getY(), event.getPos().getY(), SoundEvents.BUCKET_FILL, SoundSource.BLOCKS, 1.0F, 1.0F);
                     event.setCanceled(true);
                 }
             }
@@ -96,7 +96,7 @@ public class ForgeEventHandler
 
         if(((ItemEntity)event.getEntity()).getItem().getItem() instanceof TravelersBackpackItem)
         {
-            ((ItemEntity)event.getEntity()).setExtendedLifetime();
+            ((ItemEntity)event.getEntity()).setUnlimitedLifetime();
             event.getEntity().setInvulnerable(true);
         }
     }
@@ -104,9 +104,9 @@ public class ForgeEventHandler
     @SubscribeEvent
     public static void attachCapabilities(final AttachCapabilitiesEvent<Entity> event)
     {
-        if(event.getObject() instanceof PlayerEntity)
+        if(event.getObject() instanceof Player)
         {
-            final TravelersBackpackWearable travelersBackpack = new TravelersBackpackWearable((PlayerEntity)event.getObject());
+            final TravelersBackpackWearable travelersBackpack = new TravelersBackpackWearable((Player)event.getObject());
             event.addCapability(TravelersBackpackCapability.ID, TravelersBackpackCapability.createProvider(travelersBackpack));
         }
     }
@@ -114,9 +114,9 @@ public class ForgeEventHandler
     @SubscribeEvent
     public static void playerDeath(LivingDeathEvent event)
     {
-        if(event.getEntity() instanceof PlayerEntity)
+        if(event.getEntity() instanceof Player)
         {
-            PlayerEntity player = (PlayerEntity)event.getEntity();
+            Player player = (Player)event.getEntity();
 
             if(CapabilityUtils.isWearingBackpack(player))
             {
@@ -124,7 +124,7 @@ public class ForgeEventHandler
                 {
                     BackpackUtils.onPlayerDeath(player.level, player, CapabilityUtils.getWearingBackpack(player));
                 }
-                CapabilityUtils.synchronise((PlayerEntity)event.getEntity());
+                CapabilityUtils.synchronise((Player)event.getEntity());
             }
         }
     }
@@ -132,9 +132,12 @@ public class ForgeEventHandler
     @SubscribeEvent
     public static void playerClone(final PlayerEvent.Clone event)
     {
-        CapabilityUtils.getCapability(event.getOriginal()).ifPresent(oldTravelersBackpack ->
-                CapabilityUtils.getCapability(event.getPlayer()).ifPresent(newTravelersBackpack ->
-                        newTravelersBackpack.setWearable(oldTravelersBackpack.getWearable())));
+        Player oldPlayer = event.getOriginal();
+        oldPlayer.revive();
+
+        CapabilityUtils.getCapability(oldPlayer)
+                .ifPresent(oldTravelersBackpack -> CapabilityUtils.getCapability(event.getPlayer())
+                        .ifPresent(newTravelersBackpack -> newTravelersBackpack.setWearable(oldTravelersBackpack.getWearable())));
     }
 
     @SubscribeEvent
@@ -152,21 +155,21 @@ public class ForgeEventHandler
     @SubscribeEvent
     public static void entityJoin(EntityJoinWorldEvent event)
     {
-        if(event.getEntity() instanceof PlayerEntity)
+        if(event.getEntity() instanceof Player)
         {
-            CapabilityUtils.synchronise((PlayerEntity)event.getEntity());
+            CapabilityUtils.synchronise((Player)event.getEntity());
         }
     }
 
     @SubscribeEvent
     public static void playerTracking(final PlayerEvent.StartTracking event)
     {
-        if(event.getTarget() instanceof PlayerEntity && !event.getTarget().level.isClientSide)
+        if(event.getTarget() instanceof Player && !event.getTarget().level.isClientSide)
         {
-            ServerPlayerEntity target = (ServerPlayerEntity)event.getTarget();
+            ServerPlayer target = (ServerPlayer)event.getTarget();
 
-            CapabilityUtils.getCapability(target).ifPresent(c -> TravelersBackpack.NETWORK.send(PacketDistributor.PLAYER.with(() -> (ServerPlayerEntity)event.getPlayer()),
-                    new SyncBackpackCapabilityClient(CapabilityUtils.getWearingBackpack(target).save(new CompoundNBT()), target.getId())));
+            CapabilityUtils.getCapability(target).ifPresent(c -> TravelersBackpack.NETWORK.send(PacketDistributor.PLAYER.with(() -> (ServerPlayer) event.getPlayer()),
+                    new SyncBackpackCapabilityClient(CapabilityUtils.getWearingBackpack(target).save(new CompoundTag()), target.getId())));
         }
     }
 
@@ -192,6 +195,7 @@ public class ForgeEventHandler
             if(event.getName().equals(new ResourceLocation("chests/abandoned_mineshaft")))
             {
                 event.getTable().addPool(new LootPool.Builder().name(new ResourceLocation(TravelersBackpack.MODID, "chests/bat").toString()).build());
+                //#TODO
             }
         }
     }
@@ -205,13 +209,13 @@ public class ForgeEventHandler
         }
     }
 
-    private static class BackpackVillagerTrade implements VillagerTrades.ITrade
+    private static class BackpackVillagerTrade implements VillagerTrades.ItemListing
     {
         @Nullable
         @Override
         public MerchantOffer getOffer(Entity entity, Random random)
         {
-            return new MerchantOffer(new ItemStack(Items.EMERALD, random.nextInt(64) + 48), new ItemStack(ModItems.VILLAGER_TRAVELERS_BACKPACK.get().getItem(), 1), 1, 5, 0.5F);
+            return new MerchantOffer(new ItemStack(Items.EMERALD, random.nextInt(64) + 48), new ItemStack(ModItems.VILLAGER_TRAVELERS_BACKPACK.get().asItem(), 1), 1, 5, 0.5F);
         }
     }
 }
