@@ -1,18 +1,18 @@
 package com.tiviacz.travelersbackpack.init;
 
 import com.tiviacz.travelersbackpack.TravelersBackpack;
-import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackTileEntity;
-import net.minecraft.tileentity.TileEntityType;
-import net.minecraftforge.fml.RegistryObject;
+import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackBlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 
-public class ModTileEntityTypes {
+public class ModBlockEntityTypes {
 
-    public static final DeferredRegister<TileEntityType<?>> TILE_ENTITY_TYPES = DeferredRegister.create(ForgeRegistries.TILE_ENTITIES, TravelersBackpack.MODID);
+    public static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITY_TYPES = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITIES, TravelersBackpack.MODID);
 
-    public static final RegistryObject<TileEntityType<TravelersBackpackTileEntity>> TRAVELERS_BACKPACK = TILE_ENTITY_TYPES.register("travelers_backpack",
-            () -> TileEntityType.Builder.of(TravelersBackpackTileEntity::new,
+    public static final RegistryObject<BlockEntityType<TravelersBackpackBlockEntity>> TRAVELERS_BACKPACK = BLOCK_ENTITY_TYPES.register("travelers_backpack",
+            () -> BlockEntityType.Builder.of(TravelersBackpackBlockEntity::new,
                     ModBlocks.STANDARD_TRAVELERS_BACKPACK.get(),
                     ModBlocks.NETHERITE_TRAVELERS_BACKPACK.get(),
                     ModBlocks.DIAMOND_TRAVELERS_BACKPACK.get(),
