@@ -3,22 +3,22 @@ package com.tiviacz.travelersbackpack.util;
 import com.tiviacz.travelersbackpack.capability.CapabilityUtils;
 import com.tiviacz.travelersbackpack.capability.ITravelersBackpack;
 import com.tiviacz.travelersbackpack.config.TravelersBackpackConfig;
-import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackTileEntity;
-import net.minecraft.block.Block;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.util.Direction;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.World;
+import com.tiviacz.travelersbackpack.tileentity.TravelersBackpackBlockEntity;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.util.LazyOptional;
 
 public class BackpackUtils
 {
-    public static void onPlayerDeath(World world, PlayerEntity player, ItemStack stack)
+    public static void onPlayerDeath(Level level, Player player, ItemStack stack)
     {
         //    if(CapabilityUtils.getWearingBackpack(player).getMetadata() == 64)
          //   {
@@ -30,30 +30,31 @@ public class BackpackUtils
         {
             if(TravelersBackpackConfig.backpackForceDeathPlace)
             {
-                if(!forcePlace(world, player, stack))
+                if(!forcePlace(level, player, stack))
                 {
                     cap.ifPresent(ITravelersBackpack::removeWearable);
 
                     if(TravelersBackpackConfig.enableBackpackCoordsMessage)
                     {
-                        String translation = new TranslationTextComponent("information.travelersbackpack.backpack_drop").getString();
-                        player.sendMessage(new StringTextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
+                        String translation = new TranslatableComponent("information.travelersbackpack.backpack_drop").getString();
+                        player.sendMessage(new TextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
                         System.out.println(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ());
                     }
                 }
             }
-            else if(!tryPlace(world, player, stack))
+
+            else if(!tryPlace(level, player, stack))
             {
                 //player.entityDropItem(stack, 1); //Not too op though?
                 //int offsetY = Math.max(0, -((int) player.getPosY()) + 1) + 1;
-                player.spawnAtLocation(stack, 1);
+                player.spawnAtLocation(stack, 1); //#TODO
 
                 cap.ifPresent(ITravelersBackpack::removeWearable);
 
                 if(TravelersBackpackConfig.enableBackpackCoordsMessage)
                 {
-                    String translation = new TranslationTextComponent("information.travelersbackpack.backpack_drop").getString();
-                    player.sendMessage(new StringTextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
+                    String translation = new TranslatableComponent("information.travelersbackpack.backpack_drop").getString();
+                    player.sendMessage(new TextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
                     System.out.println(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ());
                 }
             }
@@ -61,17 +62,18 @@ public class BackpackUtils
         else
         {
             player.spawnAtLocation(stack, 1);
+
             if(TravelersBackpackConfig.enableBackpackCoordsMessage)
             {
-                String translation = new TranslationTextComponent("information.travelersbackpack.backpack_drop").getString();
-                player.sendMessage(new StringTextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
+                String translation = new TranslatableComponent("information.travelersbackpack.backpack_drop").getString();
+                player.sendMessage(new TextComponent(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ()), player.getUUID());
                 System.out.println(translation + " X: " + player.blockPosition().getX() + " Y: " + player.blockPosition().getY() + " Z: " + player.blockPosition().getZ());
             }
             cap.ifPresent(ITravelersBackpack::removeWearable);
         }
     }
 
-    public static BlockPos findBlock3D(World world, int x, int y, int z, Block block, int hRange, int vRange)
+    public static BlockPos findBlock3D(Level level, int x, int y, int z, Block block, int hRange, int vRange)
     {
         for(int i = (y - vRange); i <= (y + vRange); i++)
         {
@@ -79,7 +81,7 @@ public class BackpackUtils
             {
                 for(int k = (z - hRange); k <= (z + hRange); k++)
                 {
-                    if(world.getBlockState(new BlockPos(j, i, k)).getBlock() == block)
+                    if(level.getBlockState(new BlockPos(j, i, k)).getBlock() == block)
                     {
                         return new BlockPos(j, i, k);
                     }
@@ -89,24 +91,24 @@ public class BackpackUtils
         return null;
     }
 
-    private static boolean forcePlace(World world, PlayerEntity player, ItemStack stack)
+    private static boolean forcePlace(Level level, Player player, ItemStack stack)
     {
         if(stack.getTag() == null)
         {
-            stack.setTag(new CompoundNBT());
+            stack.setTag(new CompoundTag());
         }
 
         Block block = Block.byItem(stack.getItem());
         BlockPos playerPos = player.blockPosition();
         int y = playerPos.getY();
 
-        if(y <= 0 || y >= world.getHeight())
+        if(y <= level.getMinBuildHeight() || y >= level.getHeight())
         {
-            for(int i = 1; i < world.getHeight(); i++)
+            for(int i = 1; i < level.getHeight(); i++)
             {
                 BlockPos pos = new BlockPos(playerPos.getX(), i, playerPos.getZ());
 
-                if(world.isEmptyBlock(pos) || world.getBlockState(pos).getDestroySpeed(world, pos) > -1)
+                if(level.isEmptyBlock(pos) || level.getBlockState(pos).getDestroySpeed(level, pos) > -1)
                 {
                     y = i;
                     break;
@@ -116,26 +118,26 @@ public class BackpackUtils
 
         BlockPos targetPos = new BlockPos(playerPos.getX(), y, playerPos.getZ());
 
-        if(world.getBlockState(targetPos).getDestroySpeed(world, targetPos) > -1)
+        if(level.getBlockState(targetPos).getDestroySpeed(level, targetPos) > -1)
         {
-            while(world.getBlockEntity(targetPos) != null)
+            while(level.getBlockEntity(targetPos) != null)
             {
                 targetPos = targetPos.above();
             }
 
-            if(!world.setBlockAndUpdate(targetPos, block.defaultBlockState()))
+            if(!level.setBlockAndUpdate(targetPos, block.defaultBlockState()))
             {
                 return false;
             }
 
             if(TravelersBackpackConfig.enableBackpackCoordsMessage)
             {
-                String translation = new TranslationTextComponent("information.travelersbackpack.backpack_coords").getString();
-                player.sendMessage(new StringTextComponent(translation + " X: " + playerPos.getX() + " Y: " + y + " Z: " + playerPos.getZ()), player.getUUID());
+                String translation = new TranslatableComponent("information.travelersbackpack.backpack_coords").getString();
+                player.sendMessage(new TextComponent(translation + " X: " + playerPos.getX() + " Y: " + y + " Z: " + playerPos.getZ()), player.getUUID());
             }
 
-            world.playSound(player, playerPos.getX(), y, playerPos.getZ(), block.defaultBlockState().getSoundType().getPlaceSound(), SoundCategory.BLOCKS, 0.5F, 1.0F);
-            ((TravelersBackpackTileEntity)world.getBlockEntity(targetPos)).loadAllData(stack.getTag());
+            level.playSound(player, playerPos.getX(), y, playerPos.getZ(), block.defaultBlockState().getSoundType().getPlaceSound(), SoundSource.BLOCKS, 0.5F, 1.0F);
+            ((TravelersBackpackBlockEntity)level.getBlockEntity(targetPos)).loadAllData(stack.getTag());
 
             if(CapabilityUtils.isWearingBackpack(player))
             {
@@ -147,7 +149,7 @@ public class BackpackUtils
         return false;
     }
 
-    private static boolean tryPlace(World world, PlayerEntity player, ItemStack stack)
+    private static boolean tryPlace(Level level, Player player, ItemStack stack)
     {
         int X = (int) player.getX();
         int Z = (int) player.getZ();
@@ -155,50 +157,50 @@ public class BackpackUtils
 
         for(int Y: positions)
         {
-            BlockPos spawn = getNearestEmptyChunkCoordinatesSpiral(player, world, X, Z, new BlockPos(X, (int)player.getY() + Y, Z), 12, true, 1, (byte) 0, false);
+            BlockPos spawn = getNearestEmptyChunkCoordinatesSpiral(player, level, X, Z, new BlockPos(X, (int)player.getY() + Y, Z), 12, true, 1, (byte) 0, false);
 
             if(spawn != null)
             {
-                return placeBackpack(stack, player, world, spawn.getX(), spawn.getY(), spawn.getZ(), Direction.UP);
+                return placeBackpack(stack, player, level, spawn.getX(), spawn.getY(), spawn.getZ(), Direction.UP);
             }
         }
         return false;
     }
 
-    public static boolean placeBackpack(ItemStack stack, PlayerEntity player, World world, int x, int y, int z, Direction facing)
+    public static boolean placeBackpack(ItemStack stack, Player player, Level level, int x, int y, int z, Direction facing)
     {
         if(stack.getTag() == null)
         {
-            stack.setTag(new CompoundNBT());
+            stack.setTag(new CompoundTag());
         }
 
         Block block = Block.byItem(stack.getItem());
 
-        if(y <= 0 || y >= world.getHeight()) return false;
+        if(y <= level.getMinBuildHeight() || y >= level.getHeight()) return false;
 
         BlockPos targetPos = new BlockPos(x, y, z);
 
-        if(world.getBlockState(targetPos).getMaterial().isReplaceable())
+        if(level.getBlockState(targetPos).getMaterial().isReplaceable())
         {
-            if(!world.getBlockState(targetPos).getMaterial().isSolid())
+            if(!level.getBlockState(targetPos).getMaterial().isSolid())
             {
-                if(!world.setBlockAndUpdate(targetPos, block.defaultBlockState()))
+                if(!level.setBlock(targetPos, block.defaultBlockState(), 3))
                 {
                     return false;
                 }
 
                 if(TravelersBackpackConfig.enableBackpackCoordsMessage)
                 {
-                    String translation = new TranslationTextComponent("information.travelersbackpack.backpack_coords").getString();
-                    player.sendMessage(new StringTextComponent(translation + " X: " + x + " Y: " + y + " Z: " + z), player.getUUID());
+                    String translation = new TranslatableComponent("information.travelersbackpack.backpack_coords").getString();
+                    player.sendMessage(new TextComponent(translation + " X: " + x + " Y: " + y + " Z: " + z), player.getUUID());
                 }
 
-                world.playSound(player, x, y, z, block.defaultBlockState().getSoundType().getPlaceSound(), SoundCategory.BLOCKS, 0.5F, 1.0F);
-                ((TravelersBackpackTileEntity)world.getBlockEntity(targetPos)).loadAllData(stack.getTag());
+                level.playSound(player, x, y, z, block.defaultBlockState().getSoundType().getPlaceSound(), SoundSource.BLOCKS, 0.5F, 1.0F);
+                ((TravelersBackpackBlockEntity)level.getBlockEntity(targetPos)).loadAllData(stack.getTag());
 
                 if(stack.hasCustomHoverName())
                 {
-                    ((TravelersBackpackTileEntity) world.getBlockEntity(targetPos)).setCustomName(stack.getDisplayName());
+                    ((TravelersBackpackBlockEntity)level.getBlockEntity(targetPos)).setCustomName(stack.getHoverName());
                 }
 
                 if(CapabilityUtils.isWearingBackpack(player))
@@ -216,7 +218,7 @@ public class BackpackUtils
      * Gets you the nearest Empty Chunk Coordinates, free of charge! Looks in two dimensions and finds a block
      * that a: can have stuff placed on it and b: has space above it.
      * This is a spiral search, will begin at close range and move out.
-     * @param world  The world object.
+     * @param level  The world object.
      * @param origX  Original X coordinate
      * @param origZ  Original Z coordinate
      * @param pos      Moving X coordinate, should be the same as origX when called.
@@ -229,7 +231,7 @@ public class BackpackUtils
      * @param type   True = for player, False = for backpack
      * @return The coordinates of the block in the chunk of the world of the game of the server of the owner of the computer, where you can place something above it.
      */
-    public static BlockPos getNearestEmptyChunkCoordinatesSpiral(PlayerEntity player, World world, int origX, int origZ, BlockPos pos, int radius, boolean except, int steps, byte pass, boolean type)
+    public static BlockPos getNearestEmptyChunkCoordinatesSpiral(Player player, Level level, int origX, int origZ, BlockPos pos, int radius, boolean except, int steps, byte pass, boolean type)
     {
         //Spiral search, because Mr Darkona is awesome :)
         //This is so the backpack tries to get placed near the death point first
@@ -251,7 +253,7 @@ public class BackpackUtils
             {
                 for(; i <= pos.getX() + steps; i++)
                 {
-                    BlockPos blockPos = type ? checkCoordsForPlayer(player, world, origX, origZ, pos, except) : checkCoordsForBackpack(player, world, origX, origZ, pos, except);
+                    BlockPos blockPos = type ? checkCoordsForPlayer(player, level, origX, origZ, pos, except) : checkCoordsForBackpack(player, level, origX, origZ, pos, except);
 
                     if(blockPos != null)
                     {
@@ -259,14 +261,14 @@ public class BackpackUtils
                     }
                 }
                 pass++;
-                return getNearestEmptyChunkCoordinatesSpiral(player, world, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
+                return getNearestEmptyChunkCoordinatesSpiral(player, level, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
             }
 
             if(pass == 1)
             {
                 for(; j >= pos.getZ() - steps; j--)
                 {
-                    BlockPos blockPos = type ? checkCoordsForPlayer(player, world, origX, origZ, pos, except) : checkCoordsForBackpack(player, world, origX, origZ, pos, except);
+                    BlockPos blockPos = type ? checkCoordsForPlayer(player, level, origX, origZ, pos, except) : checkCoordsForBackpack(player, level, origX, origZ, pos, except);
 
                     if(blockPos != null)
                     {
@@ -275,7 +277,7 @@ public class BackpackUtils
                 }
                 pass--;
                 steps++;
-                return getNearestEmptyChunkCoordinatesSpiral(player, world, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
+                return getNearestEmptyChunkCoordinatesSpiral(player, level, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
             }
         }
 
@@ -285,7 +287,7 @@ public class BackpackUtils
             {
                 for(; i >= pos.getX() - steps; i--)
                 {
-                    BlockPos blockPos = type ? checkCoordsForPlayer(player, world, origX, origZ, pos, except) : checkCoordsForBackpack(player, world, origX, origZ, pos, except);
+                    BlockPos blockPos = type ? checkCoordsForPlayer(player, level, origX, origZ, pos, except) : checkCoordsForBackpack(player, level, origX, origZ, pos, except);
 
                     if(blockPos != null)
                     {
@@ -293,14 +295,14 @@ public class BackpackUtils
                     }
                 }
                 pass++;
-                return getNearestEmptyChunkCoordinatesSpiral(player, world, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
+                return getNearestEmptyChunkCoordinatesSpiral(player, level, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
             }
 
             if(pass == 1)
             {
                 for(; j <= pos.getZ() + steps; j++)
                 {
-                    BlockPos blockPos = type ? checkCoordsForPlayer(player, world, origX, origZ, pos, except) : checkCoordsForBackpack(player, world, origX, origZ, pos, except);
+                    BlockPos blockPos = type ? checkCoordsForPlayer(player, level, origX, origZ, pos, except) : checkCoordsForBackpack(player, level, origX, origZ, pos, except);
 
                     if(blockPos != null)
                     {
@@ -309,32 +311,32 @@ public class BackpackUtils
                 }
                 pass--;
                 steps++;
-                return getNearestEmptyChunkCoordinatesSpiral(player, world, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
+                return getNearestEmptyChunkCoordinatesSpiral(player, level, origX, origZ, new BlockPos(i, pos.getY(), j), radius, except, steps, pass, type);
             }
         }
         return null;
     }
 
-    private static BlockPos checkCoordsForBackpack(PlayerEntity player, World world, int origX, int origZ, BlockPos pos, boolean except)
+    private static BlockPos checkCoordsForBackpack(Player player, Level level, int origX, int origZ, BlockPos pos, boolean except)
     {
-        if(except && world.loadedAndEntityCanStandOn(pos.below(), player) && world.isEmptyBlock(pos) && !areCoordinatesTheSame(new BlockPos(origX, pos.getY(), origZ), pos))
+        if(except && level.loadedAndEntityCanStandOn(pos.below(), player) && level.isEmptyBlock(pos) && !areCoordinatesTheSame(new BlockPos(origX, pos.getY(), origZ), pos))
         {
             return pos;
         }
-        if(!except && world.loadedAndEntityCanStandOn(pos.below(), player) && world.isEmptyBlock(pos))
+        if(!except && level.loadedAndEntityCanStandOn(pos.below(), player) && level.isEmptyBlock(pos))
         {
             return pos;
         }
         return null;
     }
 
-    private static BlockPos checkCoordsForPlayer(PlayerEntity player, World world, int origX, int origZ, BlockPos pos, boolean except)
+    private static BlockPos checkCoordsForPlayer(Player player, Level level, int origX, int origZ, BlockPos pos, boolean except)
     {
-        if(except && world.loadedAndEntityCanStandOn(pos.below(), player) && world.isEmptyBlock(pos) && world.isEmptyBlock(pos.above()) && !areCoordinatesTheSame2D(origX, origZ, pos.getX(), pos.getZ()))
+        if(except && level.loadedAndEntityCanStandOn(pos.below(), player) && level.isEmptyBlock(pos) && level.isEmptyBlock(pos.above()) && !areCoordinatesTheSame2D(origX, origZ, pos.getX(), pos.getZ()))
         {
             return pos;
         }
-        if(!except && world.loadedAndEntityCanStandOn(pos.below(), player) && world.isEmptyBlock(pos) && world.isEmptyBlock(pos.above()))
+        if(!except && level.loadedAndEntityCanStandOn(pos.below(), player) && level.isEmptyBlock(pos) && level.isEmptyBlock(pos.above()))
         {
             return pos;
         }
