@@ -1,6 +1,6 @@
 package com.tiviacz.travelersbackpack.util;
 
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.items.IItemHandler;
 
 public class ItemStackUtils
